+++ src/titiler/application/tests/routes/test_tms.py
@@ -1,12 +1,18 @@
 """test TileMatrixSets endpoints."""
 
+import morecantile
+
+NB_DEFAULT_TMS = len(morecantile.tms.list())
+
 
 def test_tilematrix(app):
     """test /tileMatrixSet endpoint."""
     response = app.get("/tileMatrixSets")
     assert response.status_code == 200
     body = response.json()
-    assert len(body["tileMatrixSets"]) == 12  # morecantile has 10 defaults
+    assert (
+        len(body["tileMatrixSets"]) == NB_DEFAULT_TMS + 2
+    )  # morecantile defaults + 2 customs
     tms = list(filter(lambda m: m["id"] == "EPSG3413", body["tileMatrixSets"]))[0]
     assert tms["links"][0]["href"] == "http://testserver/tileMatrixSets/EPSG3413"